+++ CHANGELOG.md
@@ -82,6 +82,9 @@
     - 1-AddClusterDisk.ps1
     - 2-RemoveClusterDisk.ps1
   - Added links to examples from README.md.
+- Changes to xClusterPreferredOwner
+  - Refactored the unit test for this resource to use stubs and increase coverage
+    (issue #76).
 
 ### 1.6.0.0